requirements-linux.txt
PyQt5~=5.15.11
pyqt5-qt5==5.15.2
dulwich~=0.24.2
numpy~=2.3.3
requests~=2.32.5
opencv-python~=4.11.0.86
adbutils~=2.9.4
mss~=10.1.0
PyAutoGUI~=0.9.54
uiautomator2~=3.4.2
psutil~=7.1.0
tomli==2.3.0
tomli_w~=1.2.0
lxml~=6.0.2
grpcio>=1.76.0
grpcio-tools>=1.76.0
rich>=14.2.0
